Added redirection from `/` to `/ui`
`WeatherInput` created in `schemas.py`; schema added in `main.py`

Function and tool description connected with weather tool moved to file sin `tools` directory.
System prompts moved to separate file.

Numbers replaced by HTTP Status Codes delivered by FastAPI in `main.py`.

`inference` endpoints moved to separate package with all its dependencies.
Added dedicated routing.
`schemas.py` renamed to `model.py`
