--- app/deps.py
@@ -1,11 +1,0 @@
-from typing import AsyncGenerator
-from openai import AsyncOpenAI
-
-from app.config import settings
-
-
-async def get_llm_client() -> AsyncGenerator[AsyncOpenAI, None]:
-    client = AsyncOpenAI(
-        api_key=settings.llm.TOGETHER_API_KEY, base_url=settings.llm.BASE_URL
-    )
-    yield client
